Implemented TopOpeningsModal
